add test case for triple loading
