+++ integration-tests/tests/multichain/mod.rs
@@ -2,6 +2,8 @@ pub mod actions;
 
 use crate::with_multichain_nodes;
 use actions::wait_for;
+use mpc_recovery_integration_tests::env::containers::DockerClient;
+use mpc_recovery_node::{protocol::triple, test_utils};
 use test_log::test;
 
 #[test(tokio::test)]
@@ -37,14 +39,11 @@ async fn test_triples_and_presignatures() -> anyhow::Result<()> {
             let state_0 = wait_for::running_mpc(&ctx, 0).await?;
             assert_eq!(state_0.participants.len(), 3);
             wait_for::has_at_least_triples(&ctx, 2).await?;
-            // TODO: add test that checks #triples in datastore
+            // TODO: checks that the triples in datastore are consistent with local
             // for account_id in state_0.participants.keys() {
             //     let triple_storage = ctx.nodes.triple_storage(account_id.to_string()).await?;
-            //     // This errs out with
-            //     // Err(GcpError(BadRequest(Object {"error": Object {"code": Number(400), "message": String("Payload isn't valid for request."), "status": String("INVALID_ARGUMENT")}})))
             //     let load_res = triple_storage.load().await;
-            //     println!("result is: {:?}", load_res);
-            //     assert_eq!(load_res.unwrap().len(), 6);
+            //     assert!(load_res.unwrap().len() > 2);
             // }
             wait_for::has_at_least_presignatures(&ctx, 2).await?;
             Ok(())
@@ -65,4 +64,18 @@ async fn test_signature() -> anyhow::Result<()> {
         })
     })
     .await
+}
+
+#[test(tokio::test)]
+async fn test_triples_persistence_datastore_load_same_as_local() -> anyhow::Result<()> {
+    let docker_client = DockerClient::default();
+    let gcp_project_id = "test-triple-persistence";
+    let docker_network = "test-triple-persistence";
+    docker_client.create_network(docker_network).await?;
+    let datastore =
+        crate::env::containers::Datastore::run(&docker_client, docker_network, gcp_project_id)
+            .await?;
+    let datastore_url = datastore.local_address.clone();
+    test_utils::happy_triple_generation(Some(datastore_url)).await;
+    Ok(())
 }
