+++ node/Cargo.toml
@@ -50,5 +50,7 @@ near-sdk = "5.0.0-alpha.1"
 mpc-contract = { path = "../contract" }
 mpc-keys = { path = "../keys" }
 
+itertools = "0.12.0"
+
 [dev-dependencies]
 itertools = "0.12.0"
