+++ node/src/lib.rs
@@ -6,6 +6,7 @@ pub mod kdf;
 pub mod protocol;
 pub mod rpc_client;
 pub mod storage;
+pub mod test_utils;
 pub mod types;
 pub mod util;
 pub mod web;
