+++ node/src/protocol/cryptography.rs
@@ -306,7 +306,7 @@ impl CryptographicProtocol for RunningState {
         }
 
         let mut presignature_manager = self.presignature_manager.write().await;
-        if presignature_manager.my_len() < 2 && presignature_manager.potential_len() < 10  {
+        if presignature_manager.my_len() < 2 && presignature_manager.potential_len() < 10 {
             // To ensure there is no contention between different nodes we are only using triples
             // that we proposed. This way in a non-BFT environment we are guaranteed to never try
             // to use the same triple as any other node.
