+++ node/src/storage/triple_storage.rs
@@ -197,31 +197,33 @@ impl TripleNodeStorage for DataStoreTripleNodeStorage {
     }
 
     async fn load(&self) -> TripleResult<Vec<TripleData>> {
-        tracing::info!("loading triples using datastore");
+        tracing::debug!("loading triples using datastore");
         let account_id_val = DatastoreValue::from_value(self.account_id().into_value())?;
-        tracing::info!(?account_id_val, "triple storage load account_id");
-        let filter = Filter {
-            composite_filter: None,
-            property_filter: Some(PropertyFilter {
-                op: Some("Equal".to_string()),
-                property: Some(PropertyReference {
-                    name: Some("account_id".to_string()),
+        let filter = if self.datastore.is_emulator() {
+            None
+        } else {
+            Some(Filter {
+                composite_filter: None,
+                property_filter: Some(PropertyFilter {
+                    op: Some("Equal".to_string()),
+                    property: Some(PropertyReference {
+                        name: Some("account_id".to_string()),
+                    }),
+                    value: Some(account_id_val),
                 }),
-                value: Some(account_id_val),
-            }),
+            })
         };
-        let response = self
-            .datastore
-            .fetch_entities::<TripleData>(Some(filter))
-            .await?;
+        let response = self.datastore.fetch_entities::<TripleData>(filter).await?;
         let mut res: Vec<TripleData> = vec![];
         for entity_result in response {
             let entity = entity_result.entity.unwrap();
             let entity_value = entity.into_value();
             let triple_data = TripleData::from_value(entity_value).unwrap();
-            res.push(triple_data);
+            if triple_data.account_id == self.account_id() {
+                res.push(triple_data);
+            }
         }
-        tracing::info!("loading triples success");
+        tracing::debug!("loading triples success");
         Ok(res)
     }
